Add logic to check can user edit/delete restaurant and menu items or not

+++ restaurants/views.py
@@ -1,39 +1,51 @@
-from flask import Flask, render_template, url_for, redirect, request, flash, jsonify
+from flask import Flask, render_template, url_for, redirect, request, flash, jsonify, make_response
 from flask import session as login_session
-import random, string
-
-from oauth2client.client import flow_from_clientsecrets
-from oauth2client.client import FlowExchangeError
-import httplib2
-import json
-from flask import make_response
-import requests
-
+from oauth2client.client import flow_from_clientsecrets, FlowExchangeError
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from models import Base, Restaurant, MenuItem
+from models import Base, Restaurant, MenuItem, User
 from restaurants import app
+import httplib2, json, requests, random, string
+
 
 engine = create_engine('sqlite:///restaurantmenu.db')
 Base.metadata.bind = engine
-
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
 
 CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Restaurant Menu App"
 
-@app.route('/')
-@app.route('/restaurants/')
-def showRestaurants():
-    restaurants = session.query(Restaurant).all()
-    return render_template('restaurants.html', restaurants = restaurants)
+
+# User Helper Functions
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session['email'], img=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
 
 @app.route('/login')
 def showLogin():
 	state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in xrange(32))
 	login_session['state'] = state
 	return render_template('login.html', STATE=state)
+
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
@@ -102,6 +114,11 @@ def gconnect():
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
 
+    user_id = getUserID(login_session['email'])
+    if user_id is None:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
     output = ''
     output += '<h1>Welcome, '
     output += login_session['username']
@@ -112,6 +129,7 @@ def gconnect():
     flash("you are now logged in as %s" % login_session['username'])
     print "done!"
     return output
+
 
 @app.route('/gdisconnect')
 def gdisconnect():
@@ -142,6 +160,17 @@ def gdisconnect():
 		response.headers['Content-Type'] = 'application/json'
 		return response
 
+
+@app.route('/')
+@app.route('/restaurants/')
+def showRestaurants():
+    restaurants = session.query(Restaurant).all()
+    if 'username' not in login_session:
+        return render_template('publicrestaurants.html', restaurants = restaurants)
+    else:
+        return render_template('restaurants.html', restaurants = restaurants)
+
+
 @app.route('/restaurant/new/', methods=['GET', 'POST'])
 def newRestaurant():
 	if 'username' not in login_session:
@@ -160,11 +189,15 @@ def newRestaurant():
 	else:
 		return render_template('newRestaurant.html')
 
+
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods=['GET', 'POST'])
 def editRestaurant(restaurant_id):
-	if 'username' not in login_session:
-		return redirect('/login')
-	editedRestaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    editedRestaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if editedRestaurant.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. \
+                Please create your own restaurant in order to edit.')</script><body onload='myFunction()''>"
 	if request.method == 'POST':
 		if request.form['name']:
 			editedRestaurant.name = request.form['name']
@@ -185,11 +218,15 @@ def editRestaurant(restaurant_id):
 	else:
 		return render_template('editRestaurant.html', restaurant_id=restaurant_id, item=editedRestaurant)
 
+
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods=['GET', 'POST'])
 def deleteRestaurant(restaurant_id):
-	if 'username' not in login_session:
-		return redirect('/login')
-	restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    if restaurant.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. \
+                Please create your own restaurant in order to delete.')</script><body onload='myFunction()''>"
 	if request.method == 'POST':
 		session.delete(restaurant)
 		session.commit()
@@ -197,6 +234,7 @@ def deleteRestaurant(restaurant_id):
 		return redirect(url_for('showRestaurants'))
 	else:
 		return render_template('deleteRestaurant.html', restaurant_id=restaurant_id, item=restaurant)
+
 
 @app.route('/restaurants/JSON')
 def restaurantsJSON():
@@ -208,3 +246,92 @@ def restaurantsJSON():
 def restaurantJSON(restaurant_id):
     resaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
     return jsonify(Restaurant=resaurant.serialize)
+
+
+@app.route('/restaurant/<int:restaurant_id>/')
+@app.route('/restaurant/<int:restaurant_id>/menu')
+def showMenu(restaurant_id):
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    creator = getUserInfo(restaurant.user_id)
+    items = session.query(MenuItem).filter_by(restaurant_id=restaurant.id).all()
+    if 'username' not in login_session or creator != login_session['user_id']:
+        return render_template('publicmenu.html', restaurant = restaurant, items = items, creator=creator)
+    else:
+        return render_template('menu.html', restaurant = restaurant, items = items, creator=creator)
+
+
+@app.route('/restaurant/<int:restaurant_id>/menu/new/', methods=['GET', 'POST'])
+def newMenuItem(restaurant_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    if request.method == 'POST':
+        newItem = MenuItem(name=request.form['name'],
+                           description=request.form['description'],
+                           price=request.form['price'],
+                           course=request.form['course'],
+                           img=request.form['img'],
+                           restaurant_id=restaurant_id)
+        session.add(newItem)
+        session.commit()
+        flash("new menu item created!")
+        return redirect(url_for('showMenu', restaurant_id=restaurant_id))
+    else:
+        return render_template('newMenuItem.html', restaurant_id=restaurant_id)
+
+
+@app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET', 'POST'])
+def editMenuItem(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    editedItem = session.query(MenuItem).filter_by(id=menu_id).one()
+    if editedItem.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to edit this menu item. \
+                Please create your own in order to edit.')</script><body onload='myFunction()''>"
+    if request.method == 'POST':
+        if request.form['name']:
+            editedItem.name = request.form['name']
+        if request.form['description']:
+            editedItem.description=request.form['description']
+        if request.form['price']:
+            editedItem.price=request.form['price']
+        if request.form['course']:
+            editedItem.course=request.form['course']
+        if request.form['img']:
+            editedItem.img=request.form['img']
+        session.add(editedItem)
+        session.commit()
+        flash("menu item edited!")
+        return redirect(url_for('showMenu', restaurant_id=restaurant_id))
+    else:
+        return render_template('editMenuItem.html', restaurant_id=restaurant_id, menu_id=menu_id, item=editedItem)
+
+
+@app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/delete', methods=['GET', 'POST'])
+def deleteMenuItem(restaurant_id, menu_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+	item = session.query(MenuItem).filter_by(id=menu_id).one()
+    if item.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized to delete this menu item. \
+                Please create your own in order to delete.')</script><body onload='myFunction()''>"
+	if request.method == 'POST':
+		session.delete(item)
+		session.commit()
+		flash("menu item deleted!")
+		return redirect(url_for('showMenu', restaurant_id=restaurant_id))
+	else:
+		return render_template('deleteMenuItem.html', item=item, restaurant_id=restaurant_id)
+
+
+@app.route('/restaurant/<int:restaurant_id>/menu/JSON')
+def restaurantMenuJSON(restaurant_id):
+    restaurant = session.query(Restaurant).filter_by(id=restaurant_id).one()
+    items = session.query(MenuItem).filter_by(
+        restaurant_id=restaurant_id).all()
+    return jsonify(MenuItems=[i.serialize for i in items])
+
+
+@app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/JSON')
+def menuItemJSON(restaurant_id, menu_id):
+    item = session.query(MenuItem).filter_by(id=menu_id).one()
+    return jsonify(MenuItem=item.serialize)
